requirements.txt
flask==2.3.3
flask-cors==4.0.0
gdeltdoc==0.1.5
pandas==2.0.3
newspaper3k==0.2.8
nltk==3.8.1
openai==1.3.0
python-dotenv==1.0.0 